refactor: split per-os process impls
